setShopOrgetShop

+++ sky-server/src/main/java/com/sky/controller/admin/SetmealController.java
@@ -102,7 +102,7 @@ public class SetmealController {
      */
     @PostMapping("status/{status}")
     @ApiOperation("套餐起售、停售")
-    public Result startOrStop(@PathVariable Long status,Long id)
+    public Result startOrStop(@PathVariable Integer status,Long id)
     {
         log.info("套餐起售、停售：{},{}",status,id);
         setmealService.startOrStop(status,id);
